Refactored packages and started test for dominos.
