+++ src/main/java/com/seleniumTesting/dominos/LocationFinder.java
@@ -1,0 +1,17 @@
+package com.seleniumTesting.dominos;
+
+import org.openqa.selenium.By;
+
+/**
+ * Created by jerem on 5/12/2016.
+ */
+public class LocationFinder {
+    public static final String url = "http://www.dominos.com";
+
+    //textboxes
+    public static final By txtBoxCity = By.id("City");
+    public static final By txtBoxZipCode = By.id("Postal_Code");
+
+
+
+}
