+++ src/main/java/com/seleniumTesting/dominos/homePage.java
@@ -1,0 +1,15 @@
+package com.seleniumTesting.dominos;
+
+import org.openqa.selenium.By;
+
+/**
+ * Created by jerem on 5/12/2016.
+ */
+public class HomePage {
+    public static final String url = "http://www.dominos.com";
+
+    //links
+    public static final By linkLocation = By.xpath("//a[contains(@class,'qa-Cl_Loc c-site-nav-main-link-3')]");
+
+
+}
